Template HTML and CSS
HTML & CSS file for Grid layout example using named areas.  Grid areas allow a designer/developer to name areas on the grid giving them more meaning and thus easier to work with.  The template is setup with a basic 2 column layout.  Media queries are added to make it fully responsive in smaller screen sizes.  The HTML file uses a different color for each section so that they can be easily differentiated.  Class names are written in each section so that they can easily be reference to the named areas in the CSS code.